+++ app/src/main/java/com/homalab/android/compose/weather/data/datasource/WeatherDataSource.kt
@@ -3,5 +3,5 @@ package com.homalab.android.compose.weather.data.datasource
 import com.homalab.android.compose.weather.domain.entity.WeatherData
 
 interface WeatherDataSource {
-    suspend fun getCurrentWeatherData(id: Int, lat: Float, lon: Float): WeatherData
+    suspend fun getCurrentWeatherData(id: Int, lat: Double, lon: Double): WeatherData
 }
