+++ app/src/main/java/com/homalab/android/compose/weather/domain/WeatherRepository.kt
@@ -3,6 +3,6 @@ package com.homalab.android.compose.weather.domain
 import com.homalab.android.compose.weather.domain.entity.WeatherData
 
 interface WeatherRepository {
-    suspend fun getCurrentWeatherData(id: Int, lat: Float, lon: Float): WeatherData
+    suspend fun getCurrentWeatherData(id: Int, lat: Double, lon: Double): WeatherData
     suspend fun getLastWeatherData(): WeatherData
 }
